requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.3
Django==4.2.4
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
docopt==0.6.2
frozenlist==1.4.1
idna==3.6
install==1.3.5
multidict==6.0.5
paho-mqtt==2.1.0
Pillow==10.1.0
pipreqs==0.4.13
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.2.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yarg==0.1.9
yarl==1.9.4
zope.interface==6.1
